+++ jib-core/src/main/java/com/google/cloud/tools/jib/builder/steps/PullBaseImageStep.java
@@ -456,19 +456,12 @@ class PullBaseImageStep implements Callable<ImagesAndRegistryClient> {
 
     if (manifestList == null) {
       Verify.verify(manifestsAndConfigs.size() == 1);
-      ManifestTemplate manifest = manifestsAndConfigs.get(0).getManifest();
-      if (manifest instanceof V21ManifestTemplate) {
-        return Collections.singletonList(
-            JsonToImageTranslator.toImage((V21ManifestTemplate) manifest));
+      ManifestAndConfigTemplate manifestAndConfig = manifestsAndConfigs.get(0);
+      Optional<Image> cachedImage = getBaseImageIfAllLayersCached(manifestAndConfig, true);
+      if (!cachedImage.isPresent()) {
+        return Collections.emptyList();
       }
-
-      ContainerConfigurationTemplate containerConfig =
-          Verify.verifyNotNull(manifestsAndConfigs.get(0).getConfig());
-      PlatformChecker.checkManifestPlatform(buildContext, containerConfig);
-
-      return Collections.singletonList(
-          JsonToImageTranslator.toImage(
-              (BuildableManifestTemplate) Verify.verifyNotNull(manifest), containerConfig));
+      return Collections.singletonList(cachedImage.get());
     }
 
     // Manifest list cached. Identify matching platforms and check if all of them are cached.
@@ -484,13 +477,52 @@ class PullBaseImageStep implements Callable<ImagesAndRegistryClient> {
       if (!manifestAndConfigFound.isPresent()) {
         return Collections.emptyList();
       }
-
-      ManifestTemplate manifest = Verify.verifyNotNull(manifestAndConfigFound.get().getManifest());
-      ContainerConfigurationTemplate containerConfig =
-          Verify.verifyNotNull(manifestAndConfigFound.get().getConfig());
-      images.add(
-          JsonToImageTranslator.toImage((BuildableManifestTemplate) manifest, containerConfig));
+      Optional<Image> cachedImage =
+          getBaseImageIfAllLayersCached(manifestAndConfigFound.get(), false);
+      if (!cachedImage.isPresent()) {
+        return Collections.emptyList();
+      }
+      images.add(cachedImage.get());
     }
     return images.build();
+  }
+
+  /**
+   * Helper method to retrieve a base image from cache given manifest and container config. Does not
+   * return image if any layers of base image are missing in cache.
+   *
+   * @param manifestAndConfig stores an image manifest and a container config
+   * @param isSingleManifest true if base image is not a manifest list
+   * @return the single cached {@link Image} found, or {@code Optional#empty()} if the base image
+   *     not found in cache with all layers present.
+   * @throws BadContainerConfigurationFormatException if the container configuration is in a bad
+   *     format
+   * @throws PlatformNotFoundInBaseImageException if build target platform is not found in the base
+   *     image
+   * @throws LayerCountMismatchException LayerCountMismatchException if the manifest and
+   *     configuration contain conflicting layer information
+   */
+  private Optional<Image> getBaseImageIfAllLayersCached(
+      ManifestAndConfigTemplate manifestAndConfig, boolean isSingleManifest)
+      throws BadContainerConfigurationFormatException, PlatformNotFoundInBaseImageException,
+          LayerCountMismatchException {
+    Cache baseImageLayersCache = buildContext.getBaseImageLayersCache();
+    ManifestTemplate manifest = Verify.verifyNotNull(manifestAndConfig.getManifest());
+
+    // Verify all layers described in manifest are present in cache
+    if (!baseImageLayersCache.areAllLayersCached(manifest)) {
+      return Optional.empty();
+    }
+    if (manifest instanceof V21ManifestTemplate) {
+      return Optional.of(JsonToImageTranslator.toImage((V21ManifestTemplate) manifest));
+    }
+    ContainerConfigurationTemplate containerConfig =
+        Verify.verifyNotNull(manifestAndConfig.getConfig());
+    if (isSingleManifest) {
+      // If base image is not a manifest list, check and warn misconfigured platforms.
+      PlatformChecker.checkManifestPlatform(buildContext, containerConfig);
+    }
+    return Optional.of(
+        JsonToImageTranslator.toImage((BuildableManifestTemplate) manifest, containerConfig));
   }
 }